Displayable now supports float senses. Using interrupt to trigger screen mode change. Needs debouncing

+++ src/Displayable.h
@@ -4,7 +4,7 @@
 class Displayable {
   public:
     Displayable();
-    Displayable(const char*, int);
+    Displayable(const char*, float);
     
     const char* getString();
   private:
